requirement.txt
fastapi
uvicorn
psycopg2-binary
pydantic
configparser
python-dotenv
